AI-Services/chatbot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==16.0.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
ctranslate2==4.6.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.13.2
fastapi==0.128.0
faster-whisper==1.2.1
ffmpeg==1.4
filelock==3.20.2
Flask==3.1.0
flatbuffers==25.12.19
fonttools==4.55.3
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.1
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.12.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.7
llvmlite==0.46.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.0
ollama==0.6.1
onnxruntime==1.23.2
openai==2.16.0
openai-whisper==20250625
orjson==3.11.6
ormsgpack==1.12.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.4.1
protobuf==6.33.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyjokes==0.8.3
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==5.2.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.5.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.6.2
uuid_utils==0.14.0
uvicorn==0.40.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2026.1.29
zstandard==0.25.0
